Trishanku removed. DD uses values like Java lib
Unitee DD Auto does not use Setu for now. Would look into this later once rest of the critical features are in place.
