+++ arjuna-samples/workspace/arjex/tests/modules/s01unitee_engine/ep09_what_drives_me/ex05_fixed_data_class.py
@@ -30,6 +30,7 @@ class MyDataClass:
     def __iter__(self):
         return iter(range(5))
 
+
 @test_function(
     drive_with=data_class(MyDataClass)
 )
