+++ arjuna-samples/workspace/arjex/tests/modules/s01unitee_engine/ep09_what_drives_me/ex06_data_class_with_args.py
@@ -33,6 +33,7 @@ class MyDataClass:
     def __iter__(self):
         return iter(range(self.num))
 
+
 @test_function(
     drive_with=data_class(MyDataClass, 8)
 )
