--- arjuna/lib/trishanku/tpi/types/constants.py
@@ -1,4 +1,0 @@
-
-
-TRUES = {"on", "ON", "true", "TRUE", "yes", "YES", "y", "Y"}
-FALSES = {"off", "OFF", "false", "FALSE", "no", "NO", "n", "N"}
